requirements.txt
google-generativeai
edge-tts
moviepy
requests
python-dotenv
openai
pillow
